requirements.txt
GitPython==3.1.43
PyYAML==6.0.1
ollama==0.1.8
tenacity==8.2.3
typer==0.12.3
rich==13.7.1
Jinja2==3.1.6
anthropic==0.52.0
pydantic==2.7.1
pydantic-settings==2.2.1
